+++ authorization-server/src/main/java/com/springbite/authorization_server/repositories/JpaRegisteredClientRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface JpaRegisteredClientRepository extends JpaRepository<RegisteredClientEntity, Integer> {
+public interface JpaRegisteredClientRepository extends JpaRepository<RegisteredClientEntity, Long> {
     Optional<RegisteredClientEntity> findByClientId(String clientId);
 }
