+++ authorization-server/src/main/java/com/springbite/authorization_server/repositories/KeyPairRepository.java
@@ -3,5 +3,5 @@ package com.springbite.authorization_server.repositories;
 import com.springbite.authorization_server.entities.KeyPairEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KeyPairRepository extends JpaRepository<KeyPairEntity, Integer> {
+public interface KeyPairRepository extends JpaRepository<KeyPairEntity, Long> {
 }
